add tests
i've found some issues with formatted text, this could be seen in output
(which is not included now). there are some complications with parsing
newline characters, but i'll fix it later.

also script is now usable like command-line application. you can specify
the directory in which source files are located.

to make some tests you need to run
$ python parse.py tests

this will create directory tests/formatted_posts, where markdown posts
will be populated.

if directory is not specified, it assumes that script is in the source
    folder. if needed json file could not be found, it exits with code
    1.